@Airfooox 2024-12-13
Hey, first of all, thanks for the work.
I'll take a minute on the weekend and update cron-validate and also look at the issues. I'll comment then about this pr