@Airfooox 2024-12-14
hey there! I made some changes in the next branch due to some issues posted in this repo. 

Could you change your pr so you push to the next branch? 
And also include the changes I made (as far as I see it should be the addition of the allowStepping option to your code).

Thanks!