@albertodiazdorado 2024-12-20
@Airfooox I have opened a separate PR against `next`, please find it here: https://github.com/Airfooox/cron-validate/pull/306